+++ modules/types/utils.js
@@ -2,14 +2,10 @@
 
 const { string: astropack } = require('astropack');
 
-const MAX_ITEMS = 5;
 const SPECIAL = /[ `!@#%^&*()+\-=[\]{};':"\\|,.<>/?~]/;
-const nameFix = name => name.replace(new RegExp(SPECIAL, 'g'), '');
+const nameFix = name => name.replace(SPECIAL, '');
 const brackets = (sample, allowSkip) => {
-  if (allowSkip) {
-    const skip = astropack.case.isFirstLetter(sample) && !SPECIAL.test(sample);
-    if (skip) return sample;
-  }
+  if (allowSkip && astropack.case.isFirstLetter(sample) && !SPECIAL.test(sample)) return sample;
   // eslint-disable-next-line quotes
   const sep = sample.includes("'") ? '"' : "'";
   return sep + sample + sep;
@@ -24,4 +20,4 @@ const jsdoc = (meta, spacing = '') => {
   return result + spacing + ' */\n';
 };
 
-module.exports = { nameFix, brackets, MAX_ITEMS, jsdoc };
+module.exports = { nameFix, brackets, jsdoc };