+++ CHANGELOG.md
@@ -8,11 +8,14 @@
 - Release version
 -->
 
-<!-- ## [0.9.0][] - 2023-11-10
+## [0.9.0][] - 2023-11-00
 
 - Pre-release fixes
 - Documentation enhancements
-- Code quality improvements -->
+- Code quality improvements
+- New type - null
+- Not required fields now works only with undefined samples
+- Removed max items mechanism from type annotations module generator
 
 ## [0.8.0][] - 2023-11-08